Move fuse pure specific code into fuse_pure.rs
Removes some `#[cfg(not(feature = "libfuse"))]` guards.
