+++ src/channel/mount_options.rs
@@ -123,40 +123,6 @@ fn conflicts_with(option: &MountOption) -> Vec<MountOption> {
     }
 }
 
-#[derive(PartialEq)]
-#[cfg(not(feature = "libfuse"))]
-pub enum MountOptionGroup {
-    KernelOption,
-    KernelFlag,
-    Fusermount,
-}
-
-#[cfg(not(feature = "libfuse"))]
-pub fn option_group(option: &MountOption) -> MountOptionGroup {
-    match option {
-        MountOption::FSName(_) => MountOptionGroup::Fusermount,
-        MountOption::Subtype(_) => MountOptionGroup::Fusermount,
-        MountOption::CUSTOM(_) => MountOptionGroup::KernelOption,
-        MountOption::AutoUnmount => MountOptionGroup::Fusermount,
-        MountOption::AllowOther => MountOptionGroup::KernelOption,
-        MountOption::Dev => MountOptionGroup::KernelFlag,
-        MountOption::NoDev => MountOptionGroup::KernelFlag,
-        MountOption::Suid => MountOptionGroup::KernelFlag,
-        MountOption::NoSuid => MountOptionGroup::KernelFlag,
-        MountOption::RO => MountOptionGroup::KernelFlag,
-        MountOption::RW => MountOptionGroup::KernelFlag,
-        MountOption::Exec => MountOptionGroup::KernelFlag,
-        MountOption::NoExec => MountOptionGroup::KernelFlag,
-        MountOption::Atime => MountOptionGroup::KernelFlag,
-        MountOption::NoAtime => MountOptionGroup::KernelFlag,
-        MountOption::DirSync => MountOptionGroup::KernelFlag,
-        MountOption::Sync => MountOptionGroup::KernelFlag,
-        MountOption::Async => MountOptionGroup::KernelFlag,
-        MountOption::AllowRoot => MountOptionGroup::KernelOption,
-        MountOption::DefaultPermissions => MountOptionGroup::KernelOption,
-    }
-}
-
 // Format option to be passed to libfuse or kernel
 pub fn option_to_string(option: &MountOption) -> String {
     match option {
@@ -180,45 +146,6 @@ pub fn option_to_string(option: &MountOption) -> String {
         MountOption::DirSync => "dirsync".to_string(),
         MountOption::Sync => "sync".to_string(),
         MountOption::Async => "async".to_string(),
-    }
-}
-
-#[cfg(all(not(feature = "libfuse"), target_os = "linux"))]
-pub fn option_to_flag(option: &MountOption) -> libc::c_ulong {
-    match option {
-        MountOption::Dev => 0, // There is no option for dev. It's the absence of NoDev
-        MountOption::NoDev => libc::MS_NODEV,
-        MountOption::Suid => 0,
-        MountOption::NoSuid => libc::MS_NOSUID,
-        MountOption::RW => 0,
-        MountOption::RO => libc::MS_RDONLY,
-        MountOption::Exec => 0,
-        MountOption::NoExec => libc::MS_NOEXEC,
-        MountOption::Atime => 0,
-        MountOption::NoAtime => libc::MS_NOATIME,
-        MountOption::Async => 0,
-        MountOption::Sync => libc::MS_SYNCHRONOUS,
-        MountOption::DirSync => libc::MS_DIRSYNC,
-        _ => unreachable!(),
-    }
-}
-
-#[cfg(all(not(feature = "libfuse"), target_os = "macos"))]
-pub fn option_to_flag(option: &MountOption) -> libc::c_int {
-    match option {
-        MountOption::Dev => 0, // There is no option for dev. It's the absence of NoDev
-        MountOption::NoDev => libc::MNT_NODEV,
-        MountOption::Suid => 0,
-        MountOption::NoSuid => libc::MNT_NOSUID,
-        MountOption::RW => 0,
-        MountOption::RO => libc::MNT_RDONLY,
-        MountOption::Exec => 0,
-        MountOption::NoExec => libc::MNT_NOEXEC,
-        MountOption::Atime => 0,
-        MountOption::NoAtime => libc::MNT_NOATIME,
-        MountOption::Async => 0,
-        MountOption::Sync => libc::MNT_SYNCHRONOUS,
-        _ => unreachable!(),
     }
 }
